OT119-29: fixes
