+++ src/main/java/com/alkemy/ong/config/AmazonClient.java
@@ -1,0 +1,35 @@
+package com.alkemy.ong.config;
+
+import com.amazonaws.auth.AWSCredentials;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3Client;
+import javax.annotation.PostConstruct;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+@Service
+public class AmazonClient {
+
+  private AmazonS3 s3client;
+
+  @Value("${amazonProperties.endpointUrl}")
+  private String endpointUrl;
+  @Value("${amazonProperties.bucketName}")
+  private String bucketName;
+  @Value("${amazonProperties.accessKey}")
+  private String accessKey;
+  @Value("${amazonProperties.secretKey}")
+  private String secretKey;
+
+  @PostConstruct
+  private void initializeAmazon() {
+    AWSCredentials credentials = new BasicAWSCredentials(this.accessKey, this.secretKey);
+    this.s3client = new AmazonS3Client(credentials);
+  }
+
+  public String backetNameResponse() {
+    return bucketName;
+  }
+
+}
