--- src/main/java/com/alkemy/ong/service/abstraction/IDeleteFileAmazonService.java
@@ -1,7 +1,0 @@
-package com.alkemy.ong.service.abstraction;
-
-public interface IDeleteFileAmazonService {
-
-  String deleteFile(String fileUrl);
-
-}
