--- src/main/java/com/alkemy/ong/service/abstraction/IUploadFileAmazonService.java
@@ -1,9 +1,0 @@
-package com.alkemy.ong.service.abstraction;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface IUploadFileAmazonService {
-
-  String uploadFile(MultipartFile multipartFile);
-
-}
